+++ Wiwa/src/Wiwa/physics/PhysicsManager.h
@@ -126,7 +126,7 @@ namespace Wiwa {
 
 		void RayTest(const btVector3& ray_from_world, const btVector3& ray_to_world);
 
-		int RayTestWalls(const btVector3& ray_from_world, const btVector3& ray_to_world);
+		float RayTestWalls(const btVector3& ray_from_world, const btVector3& ray_to_world);
 
 	private:
 		bool m_Debug;
