+++ Wiwa/src/Wiwa/scene/SceneManager.h
@@ -82,7 +82,7 @@ namespace Wiwa {
 		static SceneId LoadScene(const char* scene_path, int flags=LOAD_DEFAULT);
 
 		// Load scene to memory by index
-		static void LoadSceneByIndex(uint32_t scene_index, int flags = LOAD_DEFAULT);
+		static void LoadSceneByIndex(size_t scene_index, int flags = LOAD_DEFAULT);
 
 		// Load scene to memory by name
 		static void LoadSceneByName(const char* scene_name, int flags = LOAD_DEFAULT);
