+++ WiwaEditor/src/Layers/Panels/UIEditorPanel.cpp
@@ -43,7 +43,7 @@ void UIEditorPanel::GetSelectedCanvas()
 	{
 		if (gm.canvas.at(i)->selected)
 		{
-			canvasSelectedID = i;
+			canvasSelectedID = (int)i;
 		}
 	}
 
@@ -63,7 +63,7 @@ void UIEditorPanel::DrawCanvasItems()
 		ImGui::SameLine();
 		if (ImGui::Button("Edit"))
 		{
-			elementSelected = i;
+			elementSelected = (int)i;
 			SetInitialValues(canvas->controls.at(i));
 		}
 		ImGui::SameLine();
